+++ packages/adapters/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./evm.js";
+export * from "./fractal.js";
 export * from "./interfaces.js";
+export * from "./ownership.js";
 export * from "./runes.js";